# services/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/orgs" : {
      "get" : {
        "tags" : [ "org-controller" ],
        "operationId" : "listMine",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OrgResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "org-controller" ],
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OrgRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrgResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users/{userId}/grant" : {
      "post" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "grantUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GrantBonusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "revokeUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/admin/users" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "users",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPageResponseAdminUserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/orgs" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "orgs",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPageResponseAdminOrgResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orgs/{orgId}" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "org",
        "parameters" : [ {
          "name" : "orgId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminOrgResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit" : {
      "get" : {
        "tags" : [ "admin-controller" ],
        "operationId" : "audit",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminPageResponseAdminAuditResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orgs/{orgId}" : {
      "delete" : {
        "tags" : [ "org-controller" ],
        "operationId" : "delete_3",
        "parameters" : [ {
          "name" : "orgId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "OrgRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        }
      },
      "OrgResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "plan" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "GrantBonusRequest" : {
        "type" : "object",
        "properties" : {
          "tier" : {
            "type" : "string"
          },
          "months" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
...
      "AdminPageResponseAdminUserResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AdminUserResponse"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "AdminUserResponse" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "email" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "ownedOrgs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "memberOrgs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "highestPlan" : {
            "type" : "string"
          }
        }
      },
...
      "AdminOrgResponse" : {
        "type" : "object",
        "properties" : {
          "orgId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "ownerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ownerEmail" : {
            "type" : "string"
          },
          "projects" : {
            "type" : "integer",
            "format" : "int32"
          },
          "members" : {
            "type" : "integer",
            "format" : "int32"
          },
          "events30d" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tierExpiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "tierReason" : {
            "type" : "string"
          },
          "tierGrantedByEmail" : {
            "type" : "string"
          }
        }
      },
      "AdminPageResponseAdminOrgResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AdminOrgResponse"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "AdminAuditResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ts" : {
            "type" : "string",
            "format" : "date-time"
          },
          "adminUser" : {
            "type" : "string",
            "format" : "uuid"
          },
          "adminEmail" : {
            "type" : "string"
          },
          "action" : {
            "type" : "string"
          },
          "targetType" : {
            "type" : "string"
          },
          "targetId" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "string"
          }
        }
      },
      "AdminPageResponseAdminAuditResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AdminAuditResponse"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}